// src/common/enums/permission.type.ts
export const PermissionTypeDetails = {
// ...
  HORAS_DE_VIGILANCIA: { designacao: "Horas De Vigilancia", sigla: "mgd_hv" },
// ...
  LANCAMENTO_PROGRAMA_UC: { designacao: "Lancamento do Programa Da Uc", sigla: "mgd_lp" },
// ...
  VALIDACAO_PROGRAMA_UC: { designacao: "Validação do Programa Da Uc", sigla: "mgdv_lp" },
// ...
  MINHAS_ASSIDUIDADES: { designacao: "Minhas Assiduidades", sigla: "msa_mas" },
// ...
} as const;